requirements.txt
PySide6>=6.5.0
PyYAML>=6.0
playwright>=1.40.0
pypdf>=4.0.0
python-dotenv>=1.0.0
